Use reusable choice component in external editor
Move the menu choice system from the external editor into a reusable
choice component.
