+++ src/components/mod.rs
@@ -1,5 +1,7 @@
+mod choice;
 mod edit;
 mod help;
 
+pub use choice::Choice;
 pub use edit::Edit;
 pub use help::Help;
