+++ src/external_editor/action.rs
@@ -1,0 +1,7 @@
+#[derive(Clone, Debug, PartialEq)]
+pub enum Action {
+	AbortRebase,
+	EditRebase,
+	RestoreAndAbortEdit,
+	UndoAndEdit,
+}
