+++ src/external_editor/external_editor_state.rs
@@ -1,0 +1,8 @@
+use anyhow::Error;
+
+#[derive(Debug)]
+pub enum ExternalEditorState {
+	Active,
+	Empty,
+	Error(Error),
+}
